Update all ternary_*.cpp files.
The changes are due to the new sequences in neon.txt and the new notor
operation.  As far as I can tell there are a lot of neutral changes in
the SSE/AVX2/AVX-512 versions which mainly just replace a couple
sequences like andnot+and with or+xor, but x86_32 and x86_64 also have
some legitimate improvements since ~x is used instead of X^-1.